# src/Mobile/ReportsApp/services/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Inspections API",
    "version": "v1"
  },
  "paths": {
...
    "/api/ReportConfiguration": {
      "post": {
        "tags": [
          "ReportConfiguration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReportConfigurationCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReportConfigurationCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddReportConfigurationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ReportConfiguration"
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumenReportConfiguration"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumenReportConfiguration"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumenReportConfiguration"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/ReportConfiguration/{id}": {
      "put": {
        "tags": [
          "ReportConfiguration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportConfigurationCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportConfigurationCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportConfigurationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ReportConfiguration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ReportConfiguration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReportConfigurationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportConfigurationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportConfigurationDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Reports/{id}": {
      "put": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReportQueryResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportQueryResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportQueryResult"
                }
              }
            }
          },
          "204": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddReportConfigurationCommand": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ReportType"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "formName": {
            "type": "string",
            "nullable": true
          },
          "remarksLabelText": {
            "type": "string",
            "nullable": true
          },
          "checksDefinition": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true,
            "readOnly": true
          },
          "signatureDefinitions": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "CheckListItemQueryResult": {
        "type": "object",
        "properties": {
          "checkListId": {
            "type": "integer",
            "format": "int32"
          },
          "checked": {
            "type": "integer",
            "format": "int32"
          },
          "editable": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "remarks": {
            "type": "string",
            "nullable": true
          },
          "required": {
            "type": "boolean"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "touched": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CheckListQueryResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "reportId": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "annotation": {
            "type": "string",
            "nullable": true
          },
          "checked": {
            "type": "boolean"
          },
          "checks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckListItemQueryResult"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DateTimeRange": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "NoteQueryResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "reportId": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "checked": {
            "type": "boolean"
          },
          "needsCheck": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ReportConfigurationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/ReportType"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "formName": {
            "type": "string",
            "nullable": true
          },
          "remarksLabelText": {
            "type": "string",
            "nullable": true
          },
          "checksDefinition": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "signatureDefinitions": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReportQueryResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "licenseNumber": {
            "type": "string",
            "nullable": true
          },
          "licenseName": {
            "type": "string",
            "nullable": true
          },
          "licenseKVA": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "licenseVolt": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "licenseAmp": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "licenseValidity": {
            "$ref": "#/components/schemas/DateTimeRange"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "formName": {
            "type": "string",
            "nullable": true
          },
          "remarksLabelText": {
            "type": "string",
            "nullable": true
          },
          "isClosed": {
            "type": "boolean"
          },
          "operationalReadingsId": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsVoltageL1N": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsVoltageL2N": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsVoltageL3N": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsVoltageL1L2": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsVoltageL1L3": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsVoltageL2L3": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsRunningLoadL1": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsRunningLoadL2": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsRunningLoadL3": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsMainBreakerAmp": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsMainBreakerPoles": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsMainBreakerCapacity": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsOverCurrentDTLA": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsOverCurrentDTLSec": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsOverCurrentIDMTLA": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsOverCurrentIDMTLTm": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsEarthFaultMA": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsEarthFaultELRA": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsEarthFaultELRSec": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsEarthFaultA": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsEarthFaultSec": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsMainBreakerRating": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsOverCurrentDirectActingEnabled": {
            "type": "boolean"
          },
          "operationalReadingsOverCurrentDirectActing": {
            "type": "integer",
            "format": "int32"
          },
          "operationalReadingsOverCurrentDTLEnabled": {
            "type": "boolean"
          },
          "operationalReadingsOverCurrentIDTMLEnabled": {
            "type": "boolean"
          },
          "operationalReadingsEarthFaultRoobEnabled": {
            "type": "boolean"
          },
          "operationalReadingsEarthFaultEIREnabled": {
            "type": "boolean"
          },
          "operationalReadingsEarthFaultEFEnabled": {
            "type": "boolean"
          },
          "signatures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SignatureQueryResult"
            },
            "nullable": true
          },
          "checkLists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckListQueryResult"
            },
            "nullable": true
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteQueryResult"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReportType": {
        "enum": [
          "Inspection"
        ],
        "type": "string"
      },
...
      "ResumenReportConfiguration": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "formName": {
            "type": "string",
            "nullable": true
          },
          "remarksLabelText": {
            "type": "string",
            "nullable": true
          },
          "definedCheckLists": {
            "type": "integer",
            "format": "int32"
          },
          "definedSignatures": {
            "type": "integer",
            "format": "int32"
          },
          "usedByReports": {
            "type": "integer",
            "format": "int32"
          },
          "inactive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SignatureQueryResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "annotation": {
            "type": "string",
            "nullable": true
          },
          "designation": {
            "type": "string",
            "nullable": true
          },
          "drawnSign": {
            "type": "string",
            "nullable": true
          },
          "principal": {
            "type": "boolean"
          },
          "remarks": {
            "type": "string",
            "nullable": true
          },
          "responsibleType": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "responsibleTypeName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "responsibleName": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "viewSign": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateReportCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "licenseNumber": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "isClosed": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UpdateReportConfigurationCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/ReportType"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "formName": {
            "type": "string",
            "nullable": true
          },
          "remarksLabelText": {
            "type": "string",
            "nullable": true
          },
          "checksDefinition": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true,
            "readOnly": true
          },
          "signatureDefinitions": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "jwt_auth": {
        "type": "http",
        "description": "Add Token in Headers",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "jwt_auth": [ ]
    }
  ]
}